feat: move the lib folder into src directory
